backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-admin-rangefilter==0.13.2
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.7
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-mock==1.12.1
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
